Elimina un asistente de un evento (sólo si eres el creador del evento)

+++ lib/views/event/event_details.dart
@@ -7,9 +7,14 @@ import 'package:provider/provider.dart';
 import '../../models/event.dart';
 import '../../models/user.dart';
 
-class EventDetailsView extends StatelessWidget {
+class EventDetailsView extends StatefulWidget {
   const EventDetailsView({super.key});
 
+  @override
+  State<EventDetailsView> createState() => _EventDetailsViewState();
+}
+
+class _EventDetailsViewState extends State<EventDetailsView> {
   @override
   Widget build(BuildContext context) {
     final selectedEvent = ModalRoute.of(context)!.settings.arguments as Event;
@@ -57,9 +62,14 @@ class EventDetailsView extends StatelessWidget {
   }
 }
 
-class _FormsColumn extends StatelessWidget {
+class _FormsColumn extends StatefulWidget {
   const _FormsColumn();
 
+  @override
+  State<_FormsColumn> createState() => _FormsColumnState();
+}
+
+class _FormsColumnState extends State<_FormsColumn> {
   @override
   Widget build(BuildContext context) {
     final selectedEvent = ModalRoute.of(context)!.settings.arguments as Event;
@@ -190,24 +200,36 @@ class _FormsColumn extends StatelessWidget {
         physics: const NeverScrollableScrollPhysics(),
         itemCount: asistentesList.length,
         itemBuilder: (BuildContext context, int index) {
-          final isCreator = creator == activeUserId;
-          return Row(
+          return FutureBuilder<User>(
+            future: userService.getUserWithUid(selectedEvent.creator),
+            builder: (context, snapshot) {
+            final isCreator = snapshot.hasData && snapshot.data!.id == activeUserId;
+            return Row(
             children: [
               Expanded(
                 child: Text(asistentesList[index]),
               ),
-              ElevatedButton(
+              if (isCreator)
+                ElevatedButton(
                 onPressed: () {
-                  eventService.deleteUserFromEvent(selectedEvent.id, index.toString());
-                  asistentesList.removeAt(index);
-                },
+                 
+                  if(asistentesList.length > 1){
+                    
+                      eventService.deleteUserFromEvent(selectedEvent.id, index.toString());
+                      setState(() {
+                        asistentesList.removeAt(index);
+                      });  
+              }
+              },
                 child: Text('Eliminar asistente'),
               ),
-            ],
-          );
-        },
-      );
-    } else {
+              ],
+            );
+          },
+        );
+      },
+    );
+  } else {
       return CircularProgressIndicator();
     }
   },
